+++ Classes/Bootstrap/Base.php
@@ -17,8 +17,6 @@ declare(strict_types=1);
 
 namespace Slavlee\CustomPackage\Bootstrap;
 
-use TYPO3\CMS\Core\Utility\GeneralUtility;
-use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;
 use Slavlee\CustomPackage\Bootstrap\Traits\ExtensionTrait;
 
 abstract class Base
@@ -29,79 +27,4 @@ abstract class Base
      * Does the main class purpose
      */
     abstract public function invoke();
-
-    /**
-     * Return the extension key in Namespace writing
-     * @return string
-     */
-    protected function getExtensionKeyAsNamespace(): string
-    {
-        return GeneralUtility::underscoredToUpperCamelCase($this->extensionKey);
-    }
-
-    /**
-     * Return the extension key in lower case without underscores
-     * @return string
-     */
-    protected function getExtensionKeyForFlexforms(): string
-    {
-        return strtolower($this->getExtensionKeyAsNamespace());
-    }
-
-    /**
-     * Return the LLL path as string
-     * @param string $key
-     * @return string
-     */
-    protected function getLLL(string $key): string
-    {
-        return 'LLL:EXT:' . $this->extensionKey . '/Resources/Private/Language/' . $key;
-    }
-
-    /**
-     * Return the path to FlexForm file
-     * @param string $filename
-     * @return string
-     */
-    protected function getFlexformPath(string $filename): string
-    {
-        return 'FILE:EXT:' . $this->extensionKey . '/Configuration/Flexforms/' . $filename;
-    }
-
-    /**
-     * Return the path for TsConfig files
-     * @param string $filename
-     * @param string $scope (Page or User)
-     * @return string
-     */
-    protected function getTsConfigPath(string $filename, string $scope = 'Page'): string
-    {
-        return 'Configuration/TSconfig/' . $scope . '/StaticFile/' . $filename;
-    }
-
-    /**
-     * Return path for Icons
-     * @param string $fileName
-     * @return string
-     */
-    protected function getIconPath(string $fileName): string
-    {
-        return 'EXT:' . $this->extensionKey . '/Resources/Public/Icons/' . $fileName;
-    }
-
-    /**
-     * Register a flexform
-     * @param string $pluginSignature
-     * @param string $fileName
-     * @return void
-     */
-    protected function registerFlexform(string $pluginSignature, string $fileName)
-    {
-        $GLOBALS['TCA']['tt_content']['types']['list']['subtypes_addlist'][$pluginSignature] = 'pi_flexform';
-
-        ExtensionManagementUtility::addPiFlexFormValue(
-            $pluginSignature,
-            $this->getFlexformPath($fileName)
-        );
-    }
 }